+++ keypairs/keypairs_test.go
@@ -14,6 +14,7 @@ import (
 
 func TestGenerateEncodeSeed(t *testing.T) {
 	defaultEntropy := "fakeRandomString"
+	errGenerateBytes := errors.New("generate bytes error")
 
 	tt := []struct {
 		name        string
@@ -27,10 +28,10 @@ func TestGenerateEncodeSeed(t *testing.T) {
 			name: "fail - generate bytes error",
 			malleate: func() interfaces.Randomizer {
 				rand := testutil.NewMockRandomizer(gomock.NewController(t))
-				rand.EXPECT().GenerateBytes(gomock.Any()).AnyTimes().Return(nil, errors.New("error"))
+				rand.EXPECT().GenerateBytes(gomock.Any()).AnyTimes().Return(nil, errGenerateBytes)
 				return rand
 			},
-			expectedErr: errors.New("error"),
+			expectedErr: errGenerateBytes,
 			algorithm:   crypto.ED25519(),
 		},
 		{
@@ -90,7 +91,7 @@ func TestGenerateEncodeSeed(t *testing.T) {
 
 			if tc.expectedErr != nil {
 				require.Empty(t, a)
-				require.Error(t, err, tc.expectedErr.Error())
+				require.ErrorIs(t, err, tc.expectedErr)
 			} else {
 				require.NoError(t, err)
 				require.Equal(t, tc.expected, a)
@@ -187,7 +188,7 @@ func TestDeriveClassicAddress(t *testing.T) {
 			actual, err := DeriveClassicAddress(tc.input)
 			if tc.expectedErr != nil {
 				require.Empty(t, actual)
-				require.Error(t, err, tc.expectedErr.Error())
+				require.ErrorIs(t, err, tc.expectedErr)
 			} else {
 				require.NoError(t, err)
 				require.Equal(t, tc.expected, actual)
@@ -296,7 +297,7 @@ func TestValidate(t *testing.T) {
 			actual, err := Validate(tc.inputMsg, tc.inputPubKey, tc.inputSig)
 			if tc.expectedErr != nil {
 				require.Zero(t, actual)
-				require.Error(t, err, tc.expectedErr.Error())
+				require.ErrorIs(t, err, tc.expectedErr)
 			} else {
 				require.NoError(t, err)
 				require.Equal(t, tc.expected, actual)
@@ -313,9 +314,9 @@ func TestDeriveNodeAddress(t *testing.T) {
 		expectedErr error
 	}{
 		{
-			name:        "fail - derive node address - node prefix not found",
-			inputPubKey: "x9KHn8NfbBsZV5q8bLfS72XyGqwFt5mgoPbcTV4c6qKiuPTAtXYk",
-			expectedErr: &addresscodec.EncodeLengthError{Instance: "NodePublicKey", Expected: addresscodec.NodePublicKeyLength, Input: 3},
+			name:        "fail - derive node address - node prefix mismatch",
+			inputPubKey: "rfZG9pC1cKF7q96TNZR264H9ykzKCxMyk44ZK8hFL8cNv1G3c8J",
+			expectedErr: addresscodec.ErrB58PrefixMismatch,
 		},
 		{
 			name:        "pass - derive correct node address from public key",
@@ -329,7 +330,7 @@ func TestDeriveNodeAddress(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			actual, err := DeriveNodeAddress(tc.inputPubKey, crypto.SECP256K1())
 			if tc.expectedErr != nil {
-				require.Error(t, err, tc.expectedErr.Error())
+				require.ErrorIs(t, err, tc.expectedErr)
 			} else {
 				require.NoError(t, err)
 				require.Equal(t, tc.expected, actual)